requirements.txt
antiorm==1.2.1
asgiref==3.5.2
backports.zoneinfo==0.2.1;python_version<"3.9"
beautifulsoup4==4.10.0
bootstrap4==0.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
db==0.1.1
defusedxml==0.7.1
Django==4.1
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==21.1
django-bootstrap5==21.1
django-crispy-forms==1.13.0
django-extensions==3.1.5
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-jsonapi==3.1.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-writable-nested==0.6.3
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
numpy==1.23.2
oauthlib==3.1.1
pandas==1.4.4
Pillow==9.2.0
psycopg2==2.9.1
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
reportlab==3.6.11
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.3.1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==6.2.0
zipp==3.6.0
